+++ src/+aod/+specification/Specification.m
@@ -1,5 +1,9 @@
 classdef (Abstract) Specification < handle & matlab.mixin.Heterogeneous
 
+    events 
+        ValidationFailed
+    end
+
     methods (Abstract)
         out = text(obj)
         setValue(obj, input)
@@ -7,6 +11,14 @@ classdef (Abstract) Specification < handle & matlab.mixin.Heterogeneous
 
     methods
         function obj = Specification()
+        end
+    end
+
+    methods (Access = protected)
+        function notifyListeners(obj, msg)
+            evtData = aod.specification.events.ValidationEvent(...
+                class(obj), msg);
+            notify(obj, 'ValidationFailed', evtData);
         end
     end
 end